requirements.txt
asgiref==3.8.1 ; 
django-humanize==0.1.2 ; 
django==5.1.3 ; 
djangorestframework==3.15.2 ; 
drf-yasg==1.21.8 ; 
gunicorn==23.0.0 ; 
humanize==4.11.0 ; 
inflection==0.5.1 ; 
packaging==24.2 ; 
pillow==11.0.0 ; 
psycopg2==2.9.10 ; 
python-decouple==3.8 ; 
pytz==2024.2 ; 
pyyaml==6.0.2 ; 
sqlparse==0.5.2 ; 
tzdata==2024.2 ;
uritemplate==4.1.1 ; 
whitenoise==6.8.2 ; 
